+++ src/app/shared/services/http/http-action.service.spec.ts
@@ -1,8 +1,8 @@
 import { instance, mock } from '@typestrong/ts-mockito'
 import { HttpClient } from '@angular/common/http'
-import { HttpErrorService } from '../http-error.service'
-import { HttpActionService } from '../http-action.service'
-import { ActionParser } from '../../../abstractions/action-parser.service'
+import { HttpErrorService } from './http-error.service'
+import { HttpActionService } from './http-action.service'
+import { ActionParser } from '../../abstractions/action-parser.service'
 
 describe('HttpActionService', () => {
   it('should be created', () => {
